AP-2894 Refactor UploadedEvidenceController
This controller was very large with lots of complicated logic to handle displaying uploaded files,
the mandatory categorisations, deleting files, validation, etc.

This change refactors that complication out into a series of services.

For the update action, the controller calls UploadedEvidence::Base service, passing in itself
as a parameter, and the Base service determines which child service should be called depending on which
button was pressed.  For the other two actions, :show and :list, the child service is called directly
from the controller.

All the controllers inherit from Base so as to have access to common methods.
