+++ app/services/uploaded_evidence/list_service.rb
@@ -1,0 +1,8 @@
+module UploadedEvidence
+  class ListService < Base
+    def call
+      attachment_type_options
+      self
+    end
+  end
+end
