+++ app/services/uploaded_evidence/populate_upload_form_service.rb
@@ -1,0 +1,11 @@
+module UploadedEvidence
+  class PopulateUploadFormService < Base
+    def call
+      populate_upload_form
+      uploaded_evidence_collection
+      @next_action = :show
+
+      self
+    end
+  end
+end
